Banner: triangles, headshot in largest web size
This will require some modification: we can shrink to 1200px width
before starting to lose the blue triforce at left.  We have to decide
whether we want 720 / 960px width layouts, or whether it simply jumps to
mobile at that point.

Part of the problem is that, right now, the headshot triangle is
right-justified while the blue triforce is left-justified.  As browser
width changes, only one of these can remain in place; the other must
move freely, and will eventually move off-screen as the browser window
shrinks.

One possible solution is to edit the banner background, creating special
versions for different screen sizes.  (This may be a good idea anyways,
as smaller PNGs will reduce load times on mobile.)